fix(bower): use latest version when conflicted
Once a bower install has been performed
updating a version and performing a subsequent install
results in a conflict.

Apply the forceLatest bower option to always perfer the latest
and not result in an error.

However, downgrading a component will still show an error
as there is no current option to force the ONE you want.